Handle both vertical and horizontal
