+++ ui/appui-react/src/appui-react/toolbar/new-toolbars/Surface.tsx
@@ -20,9 +20,14 @@ import { UiFramework } from "../../UiFramework.js";
 import { ProcessDetector } from "@itwin/core-bentley";
 import { useConditionalValue } from "../../hooks/useConditionalValue.js";
 import { SyncUiEventId } from "../../syncui/SyncUiEventDispatcher.js";
+import type { ToolbarContext } from "./Toolbar.js";
+
+type ToolbarContextProps = NonNullable<
+  React.ContextType<typeof ToolbarContext>
+>;
 
 interface SurfaceProps extends React.ComponentProps<typeof IUI_Surface> {
-  orientation: "horizontal" | "vertical";
+  orientation: ToolbarContextProps["orientation"];
 }
 
 /** @internal */
