+++ ui/appui-react/src/appui-react/toolbar/new-toolbars/ToolGroup.tsx
@@ -10,6 +10,7 @@ import "./ToolGroup.scss";
 import classnames from "classnames";
 import * as React from "react";
 import type { CommonProps } from "@itwin/core-react";
+import { assert } from "@itwin/core-bentley";
 import { OverflowButton } from "./OverflowButton.js";
 import { useOverflow } from "./useOverflow.js";
 import { getChildKey } from "../../layout/tool-settings/Docked.js";
@@ -26,9 +27,8 @@ type Child = ReturnType<typeof React.Children.toArray>[0];
 /** @internal */
 export function ToolGroup({ children, className, ...props }: ToolGroupProps) {
   const context = React.useContext(ToolbarContext);
-  const expandsTo = context?.expandsTo ?? "bottom";
-  const panelAlignment = context?.panelAlignment ?? "start";
-  const orientation = getOrientation(expandsTo);
+  assert(!!context);
+  const { panelAlignment, orientation } = context;
   const childrenArray = React.Children.toArray(children);
 
   const keyToChildMap = childrenArray.reduce<Map<string, Child>>(
@@ -92,19 +92,4 @@ export function ToolGroup({ children, className, ...props }: ToolGroupProps) {
       </Surface>
     </div>
   );
-}
-
-type ToolbarContextProps = React.ContextType<typeof ToolbarContext>;
-
-function getOrientation(
-  expandsTo: NonNullable<ToolbarContextProps>["expandsTo"]
-) {
-  switch (expandsTo) {
-    case "left":
-    case "right":
-      return "vertical";
-    case "top":
-    case "bottom":
-      return "horizontal";
-  }
 }
